backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.12,<3.13"
fastapi = "*"
sqlalchemy = "2.0.35"
asyncpg = "*"
greenlet = "*"
uvicorn = "*"
psycopg2-binary = "^2.9.9"
langchain = "^0.3.1"
langchain-openai = "^0.2.0"
pydantic-settings = "^2.5.2"
langchain-community = "^0.3.1"
langchain-postgres = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
